3.25.2020 3:30 am I have to do some more testing and I'm sure it has some kinks
3.25.2020 9:45
And of course I was correct. I rewrote the stack and queue classes which took time, and when I wrote the Stack portion, I pasted it in to the Queue section and that caused some errors; I didn't change some of the variables

3.26.2020
1:40
I made a shortened version that is more compact but i had to fix an error in my code where for the Queue class. I the remove_head function I didn't set self.tail = None if the head was also = None